separate modal body from modal
