+++ src/components/main/avatar/Avatar.jsx
@@ -1,6 +1,8 @@
 import css from './avatar.module.css';
 import { useRef, useState } from 'react';
 import AvatarModal from '../../modal/Modal';
+import { Button } from 'react-bootstrap';
+import ChangeAvatarModalBody from './ChangeAvaterModalBody';
 
 function AvatarSection({ imgUrl, setImgUrl }) {
   const [isOpenModal, setIsOpenModal] = useState(false);
@@ -12,14 +14,18 @@ function AvatarSection({ imgUrl, setImgUrl }) {
     setIsOpenModal(false);
   };
   return (
-    <section>
+    <section id="avatar">
       <div className={css['hero-img']}>
         <img src={imgUrl} ref={imgRef} />
       </div>
-      <button type="button" onClick={openModal}>
+      <Button type="button" onClick={openModal}>
         {!imgUrl ? 'choose avatar' : 'change avatar'}
-      </button>
-      {isOpenModal && <AvatarModal title="New avatar" onClose={closeModal} setImgUrl={setImgUrl} />}
+      </Button>
+      {isOpenModal && (
+        <AvatarModal title="Choose new avatar" onClose={closeModal} setImgUrl={setImgUrl}>
+          <ChangeAvatarModalBody setImgUrl={setImgUrl} />
+        </AvatarModal>
+      )}
     </section>
   );
 }
