+++ src/components/main/intro/Intro.jsx
@@ -1,0 +1,12 @@
+import { Button } from 'react-bootstrap';
+
+function Intro() {
+  return (
+    <section id="intro">
+      <p></p>
+      <Button>Change intro</Button>
+    </section>
+  );
+}
+
+export default Intro;
